refactor: getGraphQlUri to utils
